+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   ([#1084](https://github.com/nix-rust/nix/pull/1084))
 - Add `posix_fadvise`.
   ([#1089](https://github.com/nix-rust/nix/pull/1089))
+- Added `AF_VSOCK` to `AddressFamily`.
+  ([#1091](https://github.com/nix-rust/nix/pull/1091))
 
 ### Changed
 - Support for `ifaddrs` now present when building for Android.